Move client code into client directory
